@@ -5,11 +5,11 @@ import 'en_US.dart';
 import 'pt_BR.dart';
 
 class TranslationService extends Translations {
-  static final locale = Get.deviceLocale;
+  static Locale get locale => Get.deviceLocale!;
   static final fallbackLocale = Locale('en', 'US');
-  @override 
+  @override
   Map<String, Map<String, String>> get keys => {
-    'en_US': en_US,
-    'pt_BR': pt_BR,
-  };
+        'en_US': en_US,
+        'pt_BR': pt_BR,
+      };
 }
