@@ -1,48 +1,65 @@
 // To parse this JSON data, do
 //
-//     final CasesModel = CasesModelFromJson(jsonString);
+//     final welcome = welcomeFromJson(jsonString);
 
 import 'dart:convert';
 
 class CasesModel {
-  final Global global;
-  final List<Country> countries;
-  final String date;
-
   CasesModel({
-    this.global,
-    this.countries,
-    this.date,
+    required this.id,
+    required this.message,
+    required this.global,
+    required this.countries,
+    required this.date,
   });
 
-  static CasesModel fromRawJson(String str) =>
+  final String id;
+  final String message;
+  final Global global;
+  final List<Country> countries;
+  final DateTime date;
+
+  factory CasesModel.fromRawJson(String str) =>
       CasesModel.fromJson(json.decode(str) as Map<String, dynamic>);
 
   String toRawJson() => json.encode(toJson());
 
-  static CasesModel fromJson(dynamic json) => CasesModel(
-        global: json["Global"] == null
-            ? null
-            : Global.fromJson(json["Global"] as Map<String, dynamic>),
-        countries: json["Countries"] == null
-            ? null
-            : List<Country>.from(
-                (json["Countries"] as List<dynamic>)
-                    .map((x) => Country.fromJson(x as Map<String, dynamic>)),
-              ),
-        date: json["Date"] == null ? null : json["Date"] as String,
+  factory CasesModel.fromJson(Map<String, dynamic> json) => CasesModel(
+        id: json["ID"] as String,
+        message: json["Message"] as String,
+        global: Global.fromJson(json["Global"] as Map<String, dynamic>),
+        countries: List<Country>.from((json["Countries"] as Iterable).map(
+          (x) => Country.fromJson(x as Map<String, dynamic>),
+        )),
+        date: DateTime.parse(json["Date"] as String),
       );
 
   Map<String, dynamic> toJson() => {
-        "Global": global == null ? null : global.toJson(),
-        "Countries": countries == null
-            ? null
-            : List<dynamic>.from(countries.map((x) => x.toJson())),
-        "Date": date == null ? null : date,
+        "ID": id,
+        "Message": message,
+        "Global": global.toJson(),
+        "Countries": List<dynamic>.from(countries.map((x) => x.toJson())),
+        "Date": date.toIso8601String(),
       };
 }
 
 class Country {
+  Country({
+    required this.id,
+    required this.country,
+    required this.countryCode,
+    required this.slug,
+    required this.newConfirmed,
+    required this.totalConfirmed,
+    required this.newDeaths,
+    required this.totalDeaths,
+    required this.newRecovered,
+    required this.totalRecovered,
+    required this.date,
+    required this.premium,
+  });
+
+  final String id;
   final String country;
   final String countryCode;
   final String slug;
@@ -52,20 +69,8 @@ class Country {
   final int totalDeaths;
   final int newRecovered;
   final int totalRecovered;
-  final String date;
-
-  Country({
-    this.country,
-    this.countryCode,
-    this.slug,
-    this.newConfirmed,
-    this.totalConfirmed,
-    this.newDeaths,
-    this.totalDeaths,
-    this.newRecovered,
-    this.totalRecovered,
-    this.date,
-  });
+  final DateTime date;
+  final Premium premium;
 
   factory Country.fromRawJson(String str) =>
       Country.fromJson(json.decode(str) as Map<String, dynamic>);
@@ -73,56 +78,67 @@ class Country {
   String toRawJson() => json.encode(toJson());
 
   factory Country.fromJson(Map<String, dynamic> json) => Country(
-        country: json["Country"] == null ? null : json["Country"] as String,
-        countryCode:
-            json["CountryCode"] == null ? null : json["CountryCode"] as String,
-        slug: json["Slug"] == null ? null : json["Slug"] as String,
-        newConfirmed:
-            json["NewConfirmed"] == null ? null : json["NewConfirmed"] as int,
-        totalConfirmed: json["TotalConfirmed"] == null
-            ? null
-            : json["TotalConfirmed"] as int,
-        newDeaths: json["NewDeaths"] == null ? null : json["NewDeaths"] as int,
-        totalDeaths:
-            json["TotalDeaths"] == null ? null : json["TotalDeaths"] as int,
-        newRecovered:
-            json["NewRecovered"] == null ? null : json["NewRecovered"] as int,
-        totalRecovered: json["TotalRecovered"] == null
-            ? null
-            : json["TotalRecovered"] as int,
-        date: json["Date"] == null ? null : json["Date"] as String,
+        id: json["ID"] as String,
+        country: json["Country"] as String,
+        countryCode: json["CountryCode"] as String,
+        slug: json["Slug"] as String,
+        newConfirmed: json["NewConfirmed"] as int,
+        totalConfirmed: json["TotalConfirmed"] as int,
+        newDeaths: json["NewDeaths"] as int,
+        totalDeaths: json["TotalDeaths"] as int,
+        newRecovered: json["NewRecovered"] as int,
+        totalRecovered: json["TotalRecovered"] as int,
+        date: DateTime.parse(json["Date"] as String),
+        premium: Premium.fromJson(json["Premium"] as Map<String, dynamic>),
       );
 
   Map<String, dynamic> toJson() => {
-        "Country": country == null ? null : country,
-        "CountryCode": countryCode == null ? null : countryCode,
-        "Slug": slug == null ? null : slug,
-        "NewConfirmed": newConfirmed == null ? null : newConfirmed,
-        "TotalConfirmed": totalConfirmed == null ? null : totalConfirmed,
-        "NewDeaths": newDeaths == null ? null : newDeaths,
-        "TotalDeaths": totalDeaths == null ? null : totalDeaths,
-        "NewRecovered": newRecovered == null ? null : newRecovered,
-        "TotalRecovered": totalRecovered == null ? null : totalRecovered,
-        "Date": date == null ? null : date,
+        "ID": id,
+        "Country": country,
+        "CountryCode": countryCode,
+        "Slug": slug,
+        "NewConfirmed": newConfirmed,
+        "TotalConfirmed": totalConfirmed,
+        "NewDeaths": newDeaths,
+        "TotalDeaths": totalDeaths,
+        "NewRecovered": newRecovered,
+        "TotalRecovered": totalRecovered,
+        "Date": date.toIso8601String(),
+        "Premium": premium.toJson(),
       };
 }
 
+class Premium {
+  Premium();
+
+  factory Premium.fromRawJson(String str) =>
+      Premium.fromJson(json.decode(str) as Map<String, dynamic>);
+
+  String toRawJson() => json.encode(toJson());
+
+  factory Premium.fromJson(Map<String, dynamic> json) => Premium();
+
+  Map<String, dynamic> toJson() => {};
+}
+
 class Global {
+  Global({
+    required this.newConfirmed,
+    required this.totalConfirmed,
+    required this.newDeaths,
+    required this.totalDeaths,
+    required this.newRecovered,
+    required this.totalRecovered,
+    required this.date,
+  });
+
   final int newConfirmed;
   final int totalConfirmed;
   final int newDeaths;
   final int totalDeaths;
   final int newRecovered;
   final int totalRecovered;
-
-  Global({
-    this.newConfirmed,
-    this.totalConfirmed,
-    this.newDeaths,
-    this.totalDeaths,
-    this.newRecovered,
-    this.totalRecovered,
-  });
+  final DateTime date;
 
   factory Global.fromRawJson(String str) =>
       Global.fromJson(json.decode(str) as Map<String, dynamic>);
@@ -130,27 +146,22 @@ class Global {
   String toRawJson() => json.encode(toJson());
 
   factory Global.fromJson(Map<String, dynamic> json) => Global(
-        newConfirmed:
-            json["NewConfirmed"] == null ? null : json["NewConfirmed"] as int,
-        totalConfirmed: json["TotalConfirmed"] == null
-            ? null
-            : json["TotalConfirmed"] as int,
-        newDeaths: json["NewDeaths"] == null ? null : json["NewDeaths"] as int,
-        totalDeaths:
-            json["TotalDeaths"] == null ? null : json["TotalDeaths"] as int,
-        newRecovered:
-            json["NewRecovered"] == null ? null : json["NewRecovered"] as int,
-        totalRecovered: json["TotalRecovered"] == null
-            ? null
-            : json["TotalRecovered"] as int,
+        newConfirmed: json["NewConfirmed"] as int,
+        totalConfirmed: json["TotalConfirmed"] as int,
+        newDeaths: json["NewDeaths"] as int,
+        totalDeaths: json["TotalDeaths"] as int,
+        newRecovered: json["NewRecovered"] as int,
+        totalRecovered: json["TotalRecovered"] as int,
+        date: DateTime.parse(json["Date"] as String),
       );
 
   Map<String, dynamic> toJson() => {
-        "NewConfirmed": newConfirmed == null ? null : newConfirmed,
-        "TotalConfirmed": totalConfirmed == null ? null : totalConfirmed,
-        "NewDeaths": newDeaths == null ? null : newDeaths,
-        "TotalDeaths": totalDeaths == null ? null : totalDeaths,
-        "NewRecovered": newRecovered == null ? null : newRecovered,
-        "TotalRecovered": totalRecovered == null ? null : totalRecovered,
+        "NewConfirmed": newConfirmed,
+        "TotalConfirmed": totalConfirmed,
+        "NewDeaths": newDeaths,
+        "TotalDeaths": totalDeaths,
+        "NewRecovered": newRecovered,
+        "TotalRecovered": totalRecovered,
+        "Date": date.toIso8601String(),
       };
 }
