@@ -29,9 +29,9 @@ class CountryView extends GetView<HomeController> {
             ),
             body: Center(
               child: ListView.builder(
-                  itemCount: controller.state.countries.length,
+                  itemCount: controller.state!.countries.length,
                   itemBuilder: (context, index) {
-                    final country = controller.state.countries[index];
+                    final country = controller.state!.countries[index];
                     return ListTile(
                       onTap: () {
                         Get.toNamed('/home/country/details',
